% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_load.R
\name{calc_load}
\alias{calc_load}
\title{Load table for specified weight}
\usage{
calc_load(load, bar_weight = 45)
}
\arguments{
\item{load}{Total amount of weight to be lifted.}

\item{bar_weight}{Total weight of the bar. Default is 45 lbs.}
}
\value{
A dataframe of which plates are to be loaded and the quantity
}
\description{
Load table for specified weight
}
\examples{
# Athlete needs to load 315 lbs
# Requires three 45 lb plates on each side for a total of six plates
calc_load(315)
}
