% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epley_1rm.R
\name{epley_1rm}
\alias{epley_1rm}
\title{Calculate one-repition maximum using the Epley formula}
\usage{
epley_1rm(weight, reps)
}
\arguments{
\item{weight}{The total amount of weight lifted.}

\item{reps}{The number of repetitions completed.}
}
\value{
The estimated one-rep max.
}
\description{
Calculate one-repition maximum using the Epley formula
}
\examples{
# 135 lbs lifted 10 times results in an estimated one-rep max of 180 lbs
epley_1rm(weight = 135, reps = 10)
}
