% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epley_reps.R
\name{epley_reps}
\alias{epley_reps}
\title{Calculate reps based on one-rep max and total weight}
\usage{
epley_reps(max, weight)
}
\arguments{
\item{max}{One-repetition maximum of the athlete.}

\item{weight}{The total amount of weight the athlete wants to lift}
}
\value{
The estimated number of reps for a specified weight
}
\description{
Calculate reps based on one-rep max and total weight
}
\examples{
# Athlete's one-rep max is 180 lbs and desired load is 135 lbs
# Estimated number of reps is 10
epley_reps(180, 135)
}
