% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epley_weight.R
\name{epley_weight}
\alias{epley_weight}
\title{Calculate weight based on one-rep max and total reps}
\usage{
epley_weight(max, reps)
}
\arguments{
\item{max}{One-repetition maximum of the athlete.}

\item{reps}{Total number of repetitions programmed.}
}
\value{
The amount of weight to be lifted for a specified amount of reps.
}
\description{
Calculates the weight based on the one-rep max of the athlere and total repetitions programmed. Inverse of the Epley formula.
}
\examples{
# Athlete's max is 185 lbs and the number of reps programmed is 10
# Recommended load is 135 lbs
epley_weight(max = 185, reps = 10)
}
